+++ apps/frontend/src/features/tdp-lg/components/SubmittedTenderDetails.tsx
